BIM: Renamed file to avoid confusion with 0.21 BIM addon
